+++ R/directory.R
@@ -399,7 +399,10 @@ file_name <- function(x, compression = FALSE) {
 #'
 #' @param x A vector of files
 #' @param ts A single timestamp or vector of timestamps (default: `Sys.time()`)
-#' @param format A format to be applied to the times; set to `NULL` to skip formatting
+#' @param format A format to be applied to the times; set to `NULL` to skip
+#'   formatting
+#' @param sep A `character` vector of length 1 to separate the timestamp from
+#'   the file name
 #' @return The full name paths with the appended time stamp
 #' @export
 #' @examples
@@ -410,11 +413,18 @@ file_name <- function(x, compression = FALSE) {
 #' add_file_timestamp(file2)
 #'
 #' file.remove(file1, file2)
-add_file_timestamp <- function(x, ts = Sys.time(), format = "%Y-%m-%d %H%M%S") {
+add_file_timestamp <- function(x, ts = Sys.time(), format = "%Y-%m-%d %H%M%S", sep = " ") {
   if (!is.null(format)) {
     ts <- format(ts, format = format)
   }
+
   bn <- file_name(x)
   ext <- tools::file_ext(x)
-  file.path(dirname(x), paste0(bn, " ", ts, if (ext != "") ".", ext))
+
+  if (length(sep) > 1) {
+    warning("sep collapsed", call. = FALSE)
+    sep <- collapse0(sep)
+  }
+
+  file.path(dirname(x), paste0(bn, sep, ts, if (ext != "") ".", ext))
 }